datetime: fix _fromtimestamp for small-range floats
A typical timestamp is 1614803568. A CircuitPython float's granularity
at this magnitude is several minutes!

By avoiding the use of modf when t is an integer, the `datetime.now()`
classmethod is usable, i.e., it reutnrs a time that increases from second
to second instead of mostly being stuck and occasionally jumping forward.